Do day 3
This is probably some of the worst code I've written.

The parser is not even correct. It doesn't check for '()' after a 'do'
or 'don't'. It still works tho because the input I was given doesn't
have any dos and don'ts without parenthesis.

I also decided to be extra fancy and make my solution an iterator